+++ tests/test_commands.rs
@@ -327,3 +327,13 @@ fn pedantic() {
 
     test_args(args_1);
 }
+
+#[test]
+/// Check the stats command
+fn stats() {
+    let args_1 = &["stats", "-f", "tests/example_files/demo.ledger"];
+    let assert_1 = Command::cargo_bin("dinero").unwrap().args(args_1).assert();
+    let output_1 = String::from_utf8(assert_1.get_output().to_owned().stdout).unwrap();
+    assert!(output_1.lines().into_iter().count() > 3);
+    test_args(args_1);
+}