+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Changelog file for dinero-rs project, a command line application for managing fi
 ### Added
 - nicer error reporting
 - slightly better documentation
+- [```stats``` command](https://github.com/frosklis/dinero-rs/issues/96) that shows statistics about your ledger file
 ### Fixed
 - No need to [add a space before ```=``` in balance assertions](https://github.com/frosklis/dinero-rs/issues/40)
 - Correct parsing of transaction codes